feat: added more with pals and trainers, and more scaffolding
still on bug of adding the trainer to the battle before the battle. Adding it beforehand will allow us to preview a battle in advance. Also added some scaffolding for Museum, Trainers or others having a commander-like-ability that you can get. This ability comes from a deceased Monster or friend. Theres a bunch of other tweaks as well