feat: added parts of battle engine. animations. Fixed State bug
A bunch of updates in this one dawg. Spinning boxes, Question marks, a bunch of other nonsense. And I finally am building the attack phases yo. Its gettttin fun dawg. Exhilaratiiiiionz. Heeeeereeeeeeee is da boooomb sauuuuuccce. Carpenter Brut on the radddddio and the blooooood is pumpinzzz. trying to explain any of it to steve and he is like - oh idk but it is good that you have a vision.. LAZER VISION BABAY